web/requirements.txt
nicegui
fastapi
python-dotenv
aiomysql
discord.py
twitchAPI

# test dependencies
pytest
pytest-asyncio
pytest-mock
